requirements.txt
tensorflow-gpu
numpy
matplotlib
git+git://github.com/louisabraham/python3-midi@master

